# pkg/api/v2/api.gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.6.0 DO NOT EDIT.
package api

import (
	"context"
	"encoding/json"
	"fmt"
	"net/http"

	"github.com/labstack/echo/v4"
	"github.com/oapi-codegen/runtime"
	strictecho "github.com/oapi-codegen/runtime/strictmiddleware/echo"
)

// ErrorNotFound defines model for ErrorNotFound.
type ErrorNotFound struct {
	// Message A message describing the error
	Message *string `json:"message,omitempty"`
}

// Exercise defines model for Exercise.
type Exercise struct {
	// Muscles The muscles recruited by the exercise
	Muscles []string `json:"muscles"`

	// Name The name of the exercise
	Name string `json:"name"`
}

// PostExerciseJSONRequestBody defines body for PostExercise for application/json ContentType.
type PostExerciseJSONRequestBody = Exercise

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Fetch all exercises
	// (GET /exercise)
	GetExercises(ctx echo.Context) error
	// Create a new exercise
	// (POST /exercise)
	PostExercise(ctx echo.Context) error
	// Fetch an existing exercise
	// (GET /exercise/{exerciseId})
	GetExercise(ctx echo.Context, exerciseId string) error
}

// ServerInterfaceWrapper converts echo contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler ServerInterface
}

// GetExercises converts echo context to params.
func (w *ServerInterfaceWrapper) GetExercises(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetExercises(ctx)
	return err
}

// PostExercise converts echo context to params.
func (w *ServerInterfaceWrapper) PostExercise(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.PostExercise(ctx)
	return err
}

// GetExercise converts echo context to params.
func (w *ServerInterfaceWrapper) GetExercise(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "exerciseId" -------------
	var exerciseId string

	err = runtime.BindStyledParameterWithOptions("simple", "exerciseId", ctx.Param("exerciseId"), &exerciseId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: ""})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter exerciseId: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetExercise(ctx, exerciseId)
	return err
}

// This is a simple interface which specifies echo.Route addition functions which
// are present on both echo.Echo and echo.Group, since we want to allow using
// either of them for path registration
type EchoRouter interface {
	CONNECT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	DELETE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	GET(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	HEAD(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	OPTIONS(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PATCH(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	POST(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PUT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	TRACE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
}

// RegisterHandlers adds each server route to the EchoRouter.
func RegisterHandlers(router EchoRouter, si ServerInterface) {
	RegisterHandlersWithBaseURL(router, si, "")
}

// Registers handlers, and prepends BaseURL to the paths, so that the paths
// can be served under a prefix.
func RegisterHandlersWithBaseURL(router EchoRouter, si ServerInterface, baseURL string) {

	wrapper := ServerInterfaceWrapper{
		Handler: si,
	}

	router.GET(baseURL+"/exercise", wrapper.GetExercises)
	router.POST(baseURL+"/exercise", wrapper.PostExercise)
	router.GET(baseURL+"/exercise/:exerciseId", wrapper.GetExercise)

}

type GetExercisesRequestObject struct {
}

type GetExercisesResponseObject interface {
	VisitGetExercisesResponse(w http.ResponseWriter) error
}

type GetExercises200JSONResponse []Exercise

func (response GetExercises200JSONResponse) VisitGetExercisesResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type PostExerciseRequestObject struct {
	Body *PostExerciseJSONRequestBody
}

type PostExerciseResponseObject interface {
	VisitPostExerciseResponse(w http.ResponseWriter) error
}

type PostExercise201JSONResponse Exercise

func (response PostExercise201JSONResponse) VisitPostExerciseResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(201)

	return json.NewEncoder(w).Encode(response)
}

type PostExercise400Response struct {
}

func (response PostExercise400Response) VisitPostExerciseResponse(w http.ResponseWriter) error {
	w.WriteHeader(400)
	return nil
}

type GetExerciseRequestObject struct {
	ExerciseId string `json:"exerciseId"`
}

type GetExerciseResponseObject interface {
	VisitGetExerciseResponse(w http.ResponseWriter) error
}

type GetExercise200JSONResponse Exercise

func (response GetExercise200JSONResponse) VisitGetExerciseResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type GetExercise404JSONResponse ErrorNotFound

func (response GetExercise404JSONResponse) VisitGetExerciseResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(404)

	return json.NewEncoder(w).Encode(response)
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {
	// Fetch all exercises
	// (GET /exercise)
	GetExercises(ctx context.Context, request GetExercisesRequestObject) (GetExercisesResponseObject, error)
	// Create a new exercise
	// (POST /exercise)
	PostExercise(ctx context.Context, request PostExerciseRequestObject) (PostExerciseResponseObject, error)
	// Fetch an existing exercise
	// (GET /exercise/{exerciseId})
	GetExercise(ctx context.Context, request GetExerciseRequestObject) (GetExerciseResponseObject, error)
}

type StrictHandlerFunc = strictecho.StrictEchoHandlerFunc
type StrictMiddlewareFunc = strictecho.StrictEchoMiddlewareFunc

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
}

// GetExercises operation middleware
func (sh *strictHandler) GetExercises(ctx echo.Context) error {
	var request GetExercisesRequestObject

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.GetExercises(ctx.Request().Context(), request.(GetExercisesRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetExercises")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(GetExercisesResponseObject); ok {
		return validResponse.VisitGetExercisesResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}

// PostExercise operation middleware
func (sh *strictHandler) PostExercise(ctx echo.Context) error {
	var request PostExerciseRequestObject

	var body PostExerciseJSONRequestBody
	if err := ctx.Bind(&body); err != nil {
		return err
	}
	request.Body = &body

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.PostExercise(ctx.Request().Context(), request.(PostExerciseRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "PostExercise")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(PostExerciseResponseObject); ok {
		return validResponse.VisitPostExerciseResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}

// GetExercise operation middleware
func (sh *strictHandler) GetExercise(ctx echo.Context, exerciseId string) error {
	var request GetExerciseRequestObject

	request.ExerciseId = exerciseId

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.GetExercise(ctx.Request().Context(), request.(GetExerciseRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetExercise")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(GetExerciseResponseObject); ok {
		return validResponse.VisitGetExerciseResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}