# pkg/api/v2/openapi.yaml
openapi: 3.0.0

info:
  title: Exercise tracking
  description: API specification for creating exercises, workout plans and tracking sessions performed
  version: 0.0.1

servers:
  - url: "http://localhost:8080"

tags:
  - name: exercise
    description: The name of an exercise and the muscles recruited
  - name: session
    description: A full session comprising over multiple exercises tagged together with how many sets and repetitions to perform
  - name: workout
    description: A tracked workout session with the exercises performed, the sets and repetitions completed and the date of the workout
  - name: plan
    descrption: A grouping of sessions together and the days to perform them on

paths:
  /exercise:
    get:
      tags:
        - exercise
      summary: Fetch all exercises
      operationId: getExercises
      parameters: []
      requestBody:
        required: false
      responses:
        '200':
          description: A list of exercises
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Exercise'
    post:
      tags:
        - exercise
      summary: Create a new exercise
      operation: createExercise
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Exercise'
      responses:
        '201':
          description: Exercise created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exercise'
        '400':
          description: Invalid input

  /exercise/{exerciseId}:
    get:
      tags:
        - exercise
      summary: Fetch an existing exercise
      operationId: getExercise
      parameters:
        - name: exerciseId
          in: path
          required: true
          schema:
            type: string
          description: The unique identifier of the exercise to fetch
      requestBody:
        required: false
      responses:
        '200':
          description: Exercise exists and is fetched successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exercise'
        '404':
          description: Exercise not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorNotFound'

components:
  schemas:
    Exercise:
      type: object
      required:
        - name
        - muscles
      properties:
        name:
          type: string
          description: The name of the exercise
        muscles:
          type: array
          items:
            type: string
          description: The muscles recruited by the exercise
    ErrorNotFound:
      type: object
      properties:
        message:
          type: string
          description: A message describing the error